- Added .clang-tidy configuration with comprehensive checks
- Created tests/static_analysis/run_clang_tidy.sh for unified analysis
- Added GitHub Actions workflow for automated static analysis
- Fixed all clang-tidy warnings across kernel and userland code:
  * Resolved implicit widening conversions
  * Fixed parameter name consistency issues
  * Eliminated branch clone warnings
  * Optimized struct padding (vfs_file_t)
  * Corrected loop variable types
  * Added missing function declarations
  * Improved input validation in userland utilities
- Achieved 0 compilation errors and 0 real source warnings
- All recursive function warnings properly suppressed

+++ kernel/drivers/virtio_gpu.c
@@ -87,8 +87,8 @@ static int gpu_queue_init(virtio_gpu_device_t *dev, virtio_gpu_queue_t *vq,
     vq->used_phys = used_region->phys_addr;
     
     /* Initialize free descriptor list */
-    for (uint16_t i = 0; i < queue_size - 1; i++) {
-        vq->desc[i].next = i + 1;
+    for (uint32_t i = 0; i < queue_size - 1; i++) {
+        vq->desc[i].next = (uint16_t)(i + 1);
     }
     vq->desc[queue_size - 1].next = 0;
     vq->free_head = 0;
@@ -511,7 +511,7 @@ int virtio_gpu_init(uintptr_t base_addr, uint32_t irq)
     }
     
     g_gpu_device->fb_format = VIRTIO_GPU_FORMAT_B8G8R8X8_UNORM;
-    g_gpu_device->fb_size = g_gpu_device->fb_width * g_gpu_device->fb_height * 4;
+    g_gpu_device->fb_size = (size_t)g_gpu_device->fb_width * g_gpu_device->fb_height * 4;
     
     /* Allocate framebuffer */
     dma_region_t *fb_region = dma_alloc(g_gpu_device->fb_size, DMA_ZERO);
