+++ kernel/fs/ext2_file.c
@@ -22,7 +22,7 @@ static int read_block(void *device, uint32_t block_num, void *buffer, uint32_t b
     /* Read sectors */
     for (uint32_t i = 0; i < num_sectors; i++) {
         int ret = virtio_blk_read(sector + i, 
-                                  (uint8_t *)buffer + (i * 512), 1);
+                                  (uint8_t *)buffer + ((size_t)i * 512), 1);
         if (ret != 1) {
             set_errno(THUNDEROS_EIO);
             return -1;
